+++ asdf/_tests/test_types.py
@@ -251,7 +251,8 @@ flow_thing:
 """
     buff = helpers.yaml_to_asdf(yaml)
     with helpers.assert_no_warnings():
-        asdf.open(buff, ignore_version_mismatch=False, extensions=CustomFlowExtension())
+        with pytest.warns(AsdfDeprecationWarning, match=".*ignore_version_mismatch.*"):
+            asdf.open(buff, ignore_version_mismatch=False, extensions=CustomFlowExtension())
 
 
 def test_versioned_writing(monkeypatch):
@@ -414,7 +415,8 @@ undefined_data:
     # Make sure no warning occurs if explicitly ignored
     buff.seek(0)
     with helpers.assert_no_warnings():
-        afile = asdf.open(buff, ignore_unrecognized_tag=True)
+        with pytest.warns(AsdfDeprecationWarning, match=".*ignore_unrecognized_tag.*"):
+            afile = asdf.open(buff, ignore_unrecognized_tag=True)
 
 
 def test_newer_tag():