+++ asdf/treeutil.py
@@ -7,7 +7,7 @@ import warnings
 from contextlib import contextmanager
 
 from . import tagged
-from .exceptions import AsdfWarning
+from .exceptions import AsdfDeprecationWarning, AsdfWarning
 
 __all__ = ["walk", "iter_tree", "walk_and_modify", "get_children", "is_container", "PendingValue", "RemoveNode"]
 
@@ -220,7 +220,7 @@ class _RemoveNode:
 RemoveNode = _RemoveNode()
 
 
-def walk_and_modify(top, callback, ignore_implicit_conversion=False, postorder=True, _context=None):
+def walk_and_modify(top, callback, ignore_implicit_conversion=None, postorder=True, _context=None):
     """Modify a tree by walking it with a callback function.  It also has
     the effect of doing a deep copy.
 
@@ -253,6 +253,7 @@ def walk_and_modify(top, callback, ignore_implicit_conversion=False, postorder=T
         parents first.  Defaults to `True`.
 
     ignore_implicit_conversion : bool
+        DEPRECATED
         Controls whether warnings should be issued when implicitly converting a
         given type instance in the tree into a serializable object. The primary
         case for this is currently ``namedtuple``.
@@ -265,6 +266,13 @@ def walk_and_modify(top, callback, ignore_implicit_conversion=False, postorder=T
         The modified tree.
 
     """
+    if ignore_implicit_conversion is not None:
+        warnings.warn(
+            "The ignore_implicit_conversion argument to AsdfFile will be removed in 3.0.", AsdfDeprecationWarning
+        )
+    else:
+        ignore_implicit_conversion = False
+
     callback_arity = callback.__code__.co_argcount
     if callback_arity < 1 or callback_arity > 2:
         msg = "Expected callback to accept one or two arguments"
@@ -357,7 +365,11 @@ def walk_and_modify(top, callback, ignore_implicit_conversion=False, postorder=T
             # list representing the contents. Currently this is primarily
             # intended to handle namedtuple and NamedTuple instances.
             if not ignore_implicit_conversion:
-                warnings.warn(f"Failed to serialize instance of {type(node)}, converting to list instead", AsdfWarning)
+                warnings.warn(
+                    f"Failed to serialize instance of {type(node)}, converting to list instead. "
+                    "This warning will become an exception in asdf 3.0.",
+                    AsdfWarning,
+                )
             result = contents
 
         return result